Rebasing
This commit fixes issues after rebasing the most current master.
