+++ go.mod
@@ -12,7 +12,6 @@ require (
 	github.com/urfave/cli/v2 v2.23.7
 	github.com/vshn/component-appcat/apis v0.0.0-20230331085454-9d11bb2726ff
 	go.uber.org/zap v1.24.0
-	gopkg.in/yaml.v2 v2.4.0
 	k8s.io/api v0.26.3
 	k8s.io/apimachinery v0.26.3
 	k8s.io/utils v0.0.0-20230313181309-38a27ef9d749
@@ -71,6 +70,7 @@ require (
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.28.2-0.20220831092852-f930b1dc76e8 // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 	k8s.io/apiextensions-apiserver v0.26.3 // indirect
 	k8s.io/client-go v0.26.3 // indirect
